everything compiles and works, kind of.
added secret feature. 

login works, BUT doesn't compare the entered username to a list of
existing usernames. So it ends up just creating a new username every
time. Which is great for single player, but not so great for multiple
players.

The good news is that the way the rest of the methods are implemented
with currentUser as an argument, as soon as we get the login and
createNewUser methods up and running correctly it should be super easy
to finish.

Fixed some basic formatting

Started commenting javadoc